Gradient Descent Encoding (one variable)
This Python script aims to find the minimum of a mathematical function using the gradient descent method. Below is a brief description of this script:

This script performs the process of finding the minimum point of a mathematical function using the gradient descent method. The first derivative (f_derivative) is calculated and a new x value is updated in each iteration. The number of iterations and alpha value can be adjusted.